Allow masked data in coords (#6468)
* Allow masked data to be passed into a Cell, and by extensions a Coord.

* Handled case where integer array with masked values could not be created.

* Handle case where reading of scalar masked value from netCDF file
looses the original dtype.

* Removed some comments

* Removed fill_value keyword to masked_array creation - wasn't fixing the
issue with numpy warnings about NaNs.

* Single method for handling points arrays with masked data

* Refactored to add small inline helper function to add AuxCoord

* Refactored the masked array handling and now filling masked arrays with default fill type.

* Added test for Cell hashing

* test_Cell.py: Converted unittest to pytest

* Added tests for merging of scalar aux coords with mising data.

* Removed unnecessary __main__ entry point from test_Cell.py

* Missing hash test for bounded points

* More rigorous test on masked data of merged AuxCoord

* Migrated test_merge.py to pytest.

* Moved tests/test_merge.py to tests/unit/merge/test_merge.py

* What's new entry.

* Fixed some typos in comment

* Split out integration test from unit/merge/test_merge.py to
integration/merge/test_merge.py
